admission: move store{AdmissionStatsRequestEstimates}
Pure code movement. They're part of the interface definition, so better
placed in the top-level admission.go housing the remaining interfaces.

Release note: None
